backend/requirements.txt
blinker==1.8.2
click==8.1.8
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
importlib-metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==2.1.5
PyJWT==2.9.0
SQLAlchemy==2.0.41
typing-extensions==4.13.2
Werkzeug==3.0.6
zipp==3.20.2
Flask-Migrate==4.0.4
pymysql==1.0.3
